Split WindowDockRegistry into its own source file
